labeler/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.114.2
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httptools==0.6.1
idna==3.8
numpy==2.1.1
packaging==24.1
pandas==2.2.2
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.0.1
